refactor: Separate code.
