+++ sequvice_app/api/companies/v1/company/resources.py
@@ -1,9 +1,7 @@
-from flask import request
-
 from sequvice_app import app
-from sequvice_app.models import Company, SellPoint
+from sequvice_app.models import Company
 from sequvice_app.api.companies import check_auth
-from .schemas import CompanyData, SellPointData
+from .schemas import CompanyData
 
 ROOT_PATH = "/api/companies/v1/company"
 
@@ -16,22 +14,3 @@ async def get_company(company, *args, **kwargs):
     if company:
         return CompanyData().dump(company)
     return {}, 404
-
-
-@app.route(f"{ROOT_PATH}/sellpoint", methods=["POST"])
-@check_auth
-async def create_sell_point(company):
-    data = request.get_json()
-    schema = SellPointData()
-    data = schema.load(data)
-
-    sellpoint = await SellPoint.select().where(
-        SellPoint.name == data["name"],
-        SellPoint.owner == company
-    ).first()
-
-    if not sellpoint:
-        data["owner"] = company
-        sellpoint = await SellPoint.create(**data)
-
-    return schema.dump(sellpoint)
