+++ sequvice_app/api/companies/v1/company/schemas.py
@@ -13,8 +13,3 @@ class CompanyRegisteredData(CompanyData):
 class CompanyLoginData(ma.Schema):
     email = ma.fields.Email(required=True)
     password = ma.fields.String(required=True, load_only=True)
-
-
-class SellPointData(ma.Schema):
-    name = ma.fields.String(required=True)
-    logo = ma.fields.URL(required=False)
