+++ sequvice_app/api/companies/v1/company/tests.py
@@ -1,7 +1,7 @@
 import asyncio
 
 from sequvice_app.tests.helpers import autoflash
-from sequvice_app.models import Company, SellPoint
+from sequvice_app.models import Company
 
 
 @autoflash
@@ -72,57 +72,3 @@ async def test_companies(app, client):
     assert company.id == 1
     assert company.name == "Pavel"
     assert company.email == "Kus@Kus.ru"
-
-
-@autoflash
-async def test_creating_sellpoint(app, client):
-    loop = asyncio.get_running_loop()
-
-    def sync_test():
-        # register company (create new account)
-        with client.post(
-            "/api/companies/v1/company/auth/register",
-            json={"name": "Pavel", "email": "Kus@Kus.ru", "password": "MyPass"}
-        ) as res:
-            assert res.status_code == 302
-
-        # try login and get token
-        with client.post(
-            "/api/companies/v1/company/auth/login",
-            json={"email": "Kus@Kus.ru", "password": "MyPass"}
-        ) as res:
-            assert res.status_code == 200
-            json = res.get_json()
-            token = json["token"]
-
-        # create new sellpoint
-        datas = {
-            "name": "Варгаш Пахилий",
-            "logo": "https://vk.com/pahilij_avatar"
-        }
-        auth = {"Authorization": f"Bearer {token}"}
-        with client.post(
-            "/api/companies/v1/company/sellpoint",
-            headers=auth,
-            json=datas
-        ) as res:
-            assert res.status_code == 200
-            json = res.json
-            assert json
-            assert json == datas
-
-    await loop.run_in_executor(None, sync_test)
-
-    assert await Company.select().count() == 1
-    company = await Company.select().first()
-    assert company.id == 1
-    assert company.name == "Pavel"
-    assert company.email == "Kus@Kus.ru"
-    assert await company.points.count() == 1
-
-    assert await SellPoint.select().count() == 1
-    sellpoint = await SellPoint.get()
-    assert sellpoint.id == 1
-    assert sellpoint.name == "Варгаш Пахилий"
-    assert sellpoint.logo == "https://vk.com/pahilij_avatar"
-    assert await sellpoint.owner == company
