+++ sequvice_app/api/companies/v1/sellpoint/__init__.py
@@ -1,0 +1,1 @@
+from .resources import *
