+++ sequvice_app/api/companies/v1/sellpoint/schemas.py
@@ -1,0 +1,6 @@
+import marshmallow as ma
+
+
+class SellPointData(ma.Schema):
+    name = ma.fields.String(required=True)
+    logo = ma.fields.URL(required=False)
